# docs/api/swagger.yaml
basePath: /
definitions:
  github_com_zhian9_blogo-server_pkg_errors.Error:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      detail:
        description: 详细描述
        type: string
      id:
        description: 错误标识（前端友好）
        type: string
      status:
        description: 状态文本（如 "Not Found"）
        type: string
    type: object
...
  schema.Statistics:
    properties:
      created_at:
        description: 创建时间
        type: string
      date:
        description: 日期（格式：2025-10-28）
        type: string
      id:
        description: 统计ID
        type: string
      ip_count:
        description: 独立IP数
        type: integer
      pv:
        description: 页面浏览量
        type: integer
      updated_at:
        description: 更新时间
        type: string
      uv:
        description: 独立访客数
        type: integer
    type: object
  schema.Tag:
    properties:
      created_at:
        description: 创建时间
        type: string
      id:
        description: 唯一ID
        type: string
      name:
        description: 标签名称（唯一）
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  schema.TagForm:
    properties:
      name:
        description: 标签名称
        maxLength: 100
        type: string
    required:
    - name
    type: object
...
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_zhian9_blogo-server_pkg_errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
host: localhost:8040
info:
  contact:
    email: lxyaa911@gmail.com
    name: 李星云 (lxy911)
    url: https://github.com/zhian9
  description: Modern SaaS Blog Platform — REST API for content management, user system,
    RBAC, analytics, and more.
  license:
    name: MIT
    url: https://github.com/zhian9/Blogo/blob/master/LICENSE
  termsOfService: https://github.com/zhian9/Blogo
  title: Blogo API
  version: v1.0.0
paths:
...
  /api/v1/statistics:
    get:
      parameters:
      - default: 1
        description: pagination index
        in: query
        name: current
        type: integer
      - default: 10
        description: pagination size
        in: query
        name: pageSize
        type: integer
      - description: Date >= (YYYY-MM-DD)
        in: query
        name: date_gte
        type: string
      - description: Date <= (YYYY-MM-DD)
        in: query
        name: date_lte
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Statistics'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Query statistics list
      tags:
      - StatisticsAPI
  /api/v1/statistics/{date}:
    get:
      parameters:
      - description: Date (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Statistics'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Get statistics by date
      tags:
      - StatisticsAPI
  /api/v1/statistics/latest:
    get:
      parameters:
      - default: 7
        description: Number of days
        in: query
        name: days
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Statistics'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      summary: Get latest N days statistics (public)
      tags:
      - StatisticsAPI
  /api/v1/tags:
    get:
      parameters:
      - default: 1
        description: pagination index
        in: query
        name: current
        type: integer
      - default: 10
        description: pagination size
        in: query
        name: pageSize
        type: integer
      - description: Tag name (fuzzy)
        in: query
        name: name
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Tag'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Query tag list
      tags:
      - TagAPI
    post:
      parameters:
      - description: Tag form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.TagForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Tag'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Create tag
      tags:
      - TagAPI
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Delete tag
      tags:
      - TagAPI
    get:
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Tag'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Get tag by ID
      tags:
      - TagAPI
    put:
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      - description: Tag form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.TagForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Update tag
      tags:
      - TagAPI
  /api/v1/tags/all:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Tag'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      summary: Get all tags (public, for tag cloud)
      tags:
      - TagAPI
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your JWT token>". Login via /api/v1/login to obtain
      a token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"